requirements.txt
torch
torchvision
timm
streamlit
pillow
numpy
tqdm
requests
